+++ Craftsman/Helpers/EntityScaffolding.cs
@@ -105,12 +105,12 @@
                     feature, projectBaseName);
             }
             
-            if (feature.Type == FeatureType.AddListforFk.Name)
+            if (feature.Type == FeatureType.AddListByFk.Name)
             {
                 CommandAddListBuilder.CreateCommand(srcDirectory, entity, dbContextName, projectBaseName, feature, fileSystem);
                 AddListCommandTestBuilder.CreateTests(testDirectory, entity, feature, projectBaseName, fileSystem);
                 AddListTestBuilder.CreateTests(testDirectory, entity, policies, feature, projectBaseName, fileSystem);
-                ControllerModifier.AddEndpoint(srcDirectory, FeatureType.AddListforFk, entity, addSwaggerComments, policies, 
+                ControllerModifier.AddEndpoint(srcDirectory, FeatureType.AddListByFk, entity, addSwaggerComments, policies, 
                     feature, projectBaseName);
             }