+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Added BaseEntity that all entities will inherit from.
   * TODO: docs
 * Added built in features to the `add:feature` command
+* New `AddListByFk` option for the `add:feature`  command and `Feature` property of an entity.
 
 ### Updated